MAINT: Refactor plotting functions
